refactor!: move ReturnCode into xknx.telegram.apci
xknx.management imports from xknx.telegram, so ReturnCode living in
xknx.management.application_layer_enum forced FunctionPropertyExtStateResponse
(in apci.py) into a TYPE_CHECKING/lazy-import workaround to avoid a
circular import. Moving it next to its actual users removes the need
for that entirely.

BREAKING CHANGE: xknx.management.application_layer_enum is removed.
Import ReturnCode from xknx.telegram.apci instead.
